changed structure of tests and removed creating test folder in home directory. adjusted the error messages, which the user recieves when passing invalid output size and output region
